Using a compatible version of avatica for building.
The current main of avatica fails with calcite v1.30. We should bring
back testing with avatica main once the current main is synced with
v1.37

Using a compatible version of avatica for building.

The current main of avatica fails with calcite v1.30. We should bring
back testing with avatica main once the current main is synced with
v1.37

Currently disabling the action since the testcontainers library
is pulling in a deprecated version of a redis container